+++ lib/bower-rails/railtie.rb
@@ -9,8 +9,7 @@ module BowerRails
 
     if File.exist?(@@bowerfile)
       config.before_initialize do |app|
-        @dsl = BowerRails::Dsl.evalute(@@bowerfile)
-
+        @dsl = BowerRails::Dsl.evalute(BowerRails.root_path, @@bowerfile)
         @dsl.final_assets_path.map do |assets_root, assets_path|
           app.config.assets.paths << Rails.root.join(assets_root, assets_path, "bower_components")
         end