+++ lib/generators/bower_rails/initialize/templates/bower_rails.rb
@@ -1,10 +1,13 @@
 BowerRails.configure do |bower_rails|
-  # Invoke `rake bower:install` automatically before the `rake assets:precompile` task.
+  # Tell bower-rails what path should be considered as root. Defaults to Dir.pwd
+  # bower_rails.root_path = Dir.pwd
+
+  # Invokes rake bower:install before precompilation. Defaults to false
   # bower_rails.install_before_precompile = true
 
-  # Invoke `rake bower:resolve` automatically before the `rake assets:precompile` task.
+  # Invokes rake bower:resolve before precompilation. Defaults to false
   # bower_rails.resolve_before_precompile = true
 
-  # Invoke `rake bower:clean` automatically before the `rake assets:precompile` task.
+  # Invokes rake bower:clean before precompilation. Defaults to false
   # bower_rails.clean_before_precompile = true
 end